requirements.txt
allure-pytest~=2.13.5
allure-python-commons~=2.13.5
atomicwrites==1.4.0
attrs~=23.2.0
certifi~=2024.7.4
cffi~=1.16.0
charset-normalizer~=3.3.2
colorama~=0.4.6
colorlog~=6.8.2
cryptography==36.0.0
DingtalkChatbot==1.5.3
execnet==1.9.0
Faker~=26.3.0
idna~=3.7
iniconfig~=2.0.0
jsonpath~=0.82.2
jsonpath_ng~=1.6.1
pluggy~=1.5.0
py==1.11.0
pycparser~=2.22
pydantic~=2.8.2
PyMySQL==1.0.2
pyOpenSSL==21.0.0
pyparsing==3.0.6
pytest~=8.2.2
pytest-forked==1.3.0
pytest-xdist==2.4.0
python-dateutil==2.8.2
PyYAML~=6.0.2
requests~=2.32.3
requests-toolbelt~=1.0.0
six==1.16.0
text-unidecode==1.3
toml==0.10.2
urllib3~=2.2.2
xlrd==2.0.1
xlutils==2.0.0
xlwt==1.3.0
python-dotenv==1.0.1
